+++ packages/bluesky/lib/src/entities/actor_viewer.dart
@@ -8,6 +8,8 @@
 import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import 'list_view_basic.dart';
+
 part 'actor_viewer.freezed.dart';
 part 'actor_viewer.g.dart';
 
@@ -17,6 +19,7 @@ class ActorViewer with _$ActorViewer {
   const factory ActorViewer({
     @JsonKey(name: 'muted') required bool isMuted,
     @JsonKey(name: 'blockedBy') required bool isBlockedBy,
+    ListViewBasic? mutedByList,
     @AtUriConverter() AtUri? blocking,
     @AtUriConverter() AtUri? following,
     @AtUriConverter() AtUri? followedBy,